+++ src/main/java/tf/tailfriend/global/config/OAuth2AttributeExtractor.java
@@ -1,4 +1,4 @@
-package tf.tailfriend.user.config;
+package tf.tailfriend.global.config;
 
 import java.util.Map;
