+++ src/main/java/tf/tailfriend/user/controller/AuthController.java
@@ -9,8 +9,8 @@ import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import tf.tailfriend.user.config.JwtTokenProvider;
-import tf.tailfriend.user.config.UserPrincipal;
+import tf.tailfriend.global.config.JwtTokenProvider;
+import tf.tailfriend.global.config.UserPrincipal;
 import tf.tailfriend.user.entity.User;
 import tf.tailfriend.user.entity.dto.LoginRequestDto;
 import tf.tailfriend.user.entity.dto.UserRegisterDto;
