+++ src/main/java/tf/tailfriend/user/service/AuthService.java
@@ -2,7 +2,7 @@ package tf.tailfriend.user.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import tf.tailfriend.user.config.JwtTokenProvider;
+import tf.tailfriend.global.config.JwtTokenProvider;
 import tf.tailfriend.user.entity.User;
 import tf.tailfriend.user.entity.dto.LoginRequestDto;
 import tf.tailfriend.user.repository.UserRepository;
